+++ src/My.php
@@ -7,11 +7,10 @@ namespace Dotclear\Plugin\DotclearWatch;
 use Dotclear\Module\MyPlugin;
 
 /**
- * @brief   DotclearWatch My helper.
- * @ingroup DotclearWatch
+ * @brief       DotclearWatch My helper.
+ * @ingroup     DotclearWatch
  *
  * @author      Jean-Christian Denis
- * @copyright   Jean-Christian Denis
  * @copyright   GPL-2.0 https://www.gnu.org/licenses/gpl-2.0.html
  */
 class My extends MyPlugin